+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file is used to list changes made in each version of the haproxy cookbook.
 * Added Chefspec Matchers for the resources defined in this cookbook.
 * Added `mode` property to `backend` and `frontend` resources.
 * Added `maxconn` to `global` resource
+* removed `default_backend` as a required property on the `frontend` resource
 
 ## [v4.2.0] (04-05-2017)